pyproject.toml
[tool.poetry]
name = "weather_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "weather_app"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
dadata = "^21.10.1"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
